Cygwin: pty: Use GetTickCount64() instead of GetTickCount().
Suggested-by: Corinna Vinschen <corinna@vinschen.de>

+++ winsup/cygwin/local_includes/tty.h
@@ -119,7 +119,7 @@ private:
   bool switch_to_nat_pipe;
   DWORD nat_pipe_owner_pid;
   UINT term_code_page;
-  DWORD fwd_last_time;
+  ULONGLONG fwd_last_time;
   bool fwd_not_empty;
   HANDLE h_pcon_write_pipe;
   HANDLE h_pcon_condrv_reference;
